*部分源文件改成了UTF-8格式 *修改了vs2015 clang编译错误，还没有全部清除

+++ include/GrapX/gvNode.h
@@ -44,7 +44,7 @@ enum GVNodeFlags
   GVNF_USER_4         = 0x10000000,
   GVNF_USER_5         = 0x20000000,
   GVNF_USER_6         = 0x40000000,
-  GVNF_USER_7         = 0x80000000,
+  GVNF_USER_7         = (int)0x80000000,
 
   GVNF_DEFAULT        = GVNF_VISIBLE|GVNF_CASTSHADOW,
 };
